refactor to _render
