+++ __tests__/e2e/.vitepress/config.ts
@@ -91,8 +91,11 @@ export default defineConfig({
     search: {
       provider: 'local',
       options: {
-        exclude(relativePath) {
-          return relativePath.startsWith('local-search/excluded')
+        _render(src, env, md) {
+          const html = md.render(src, env)
+          if (env.frontmatter?.search === false) return ''
+          if (env.relativePath.startsWith('local-search/excluded')) return ''
+          return html
         }
       }
     }
