+++ src/node/markdown/plugins/link.ts
@@ -3,9 +3,13 @@
 // 2. normalize internal links to end with `.html`
 
 import type MarkdownIt from 'markdown-it'
-import type { MarkdownEnv } from '../env'
 import { URL } from 'url'
-import { EXTERNAL_URL_RE, PATHNAME_PROTOCOL_RE, isExternal } from '../../shared'
+import {
+  EXTERNAL_URL_RE,
+  PATHNAME_PROTOCOL_RE,
+  isExternal,
+  type MarkdownEnv
+} from '../../shared'
 
 const indexRE = /(^|.*\/)index.md(#?.*)$/i
