+++ src/shared/shared.ts
@@ -7,10 +7,11 @@ export type {
   Header,
   LocaleConfig,
   LocaleSpecificConfig,
+  MarkdownEnv,
   PageData,
   PageDataPayload,
-  SiteData,
-  SSGContext
+  SSGContext,
+  SiteData
 } from '../../types/shared'
 
 export const EXTERNAL_URL_RE = /^[a-z]+:/i
