+++ types/shared.d.ts
@@ -1,4 +1,5 @@
 // types shared between server and client
+import type { MarkdownSfcBlocks } from '@mdit-vue/plugin-sfc'
 import type { UseDarkOptions } from '@vueuse/core'
 import type { SSRContext } from 'vue/server-renderer'
 export type { DefaultTheme } from './default-theme.js'
@@ -98,3 +99,41 @@ export type LocaleConfig<ThemeConfig = any> = Record<
   string,
   LocaleSpecificConfig<ThemeConfig> & { label: string; link?: string }
 >
+
+// Manually declaring all properties as rollup-plugin-dts
+// is unable to merge augmented module declarations
+
+export interface MarkdownEnv {
+  /**
+   * The raw Markdown content without frontmatter
+   */
+  content?: string
+  /**
+   * The excerpt that extracted by `@mdit-vue/plugin-frontmatter`
+   *
+   * - Would be the rendered HTML when `renderExcerpt` is enabled
+   * - Would be the raw Markdown when `renderExcerpt` is disabled
+   */
+  excerpt?: string
+  /**
+   * The frontmatter that extracted by `@mdit-vue/plugin-frontmatter`
+   */
+  frontmatter?: Record<string, unknown>
+  /**
+   * The headers that extracted by `@mdit-vue/plugin-headers`
+   */
+  headers?: Header[]
+  /**
+   * SFC blocks that extracted by `@mdit-vue/plugin-sfc`
+   */
+  sfcBlocks?: MarkdownSfcBlocks
+  /**
+   * The title that extracted by `@mdit-vue/plugin-title`
+   */
+  title?: string
+  path: string
+  relativePath: string
+  cleanUrls: boolean
+  links?: string[]
+  includes?: string[]
+}
